fix: MSVC 14.42 ICE on rml_sequencer_panel_timeline.cpp
Split rml_sequencer_panel_timeline.cpp into two translation units and
replace C++20 designated initializers with member assignment in
updateTimelineGuides().

MSVC 14.42.34433 (latest VS 2022 Community) crashes with fatal error
C1001 (msc1.cpp line 1599) when compiling the original 1025-line file.
The root cause is a combination of:

1. Accumulated compiler state from the large translation unit
2. C++20 designated initializers inside a lambda returning
   std::optional<struct> — this specific pattern triggers the ICE

CI passes because GitHub windows-2022 runners use VS 2022 Enterprise
with a newer/patched MSVC toolchain.

Verified: clean build on MSVC 14.42.34433, Windows 11, CUDA 12.9.
